Replace half_join with half_join2
