+++ dogsdogsdogs/src/operators/mod.rs
@@ -1,5 +1,4 @@
 pub mod half_join;
-pub mod half_join2;
 pub mod lookup_map;
 
 pub mod count;
@@ -7,7 +6,6 @@ pub mod propose;
 pub mod validate;
 
 pub use self::half_join::half_join;
-pub use self::half_join2::half_join as half_join2;
 pub use self::lookup_map::lookup_map;
 pub use self::count::count;
 pub use self::propose::{propose, propose_distinct};
